+++ app/src/main/java/woowacourse/shopping/feature/cart/adapter/CartViewHolder.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.feature.cart.adapter
 
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import woowacourse.shopping.databinding.ItemCartBinding
-import woowacourse.shopping.feature.main.QuantityControlListener
+import woowacourse.shopping.common.QuantityControlListener
 import woowacourse.shopping.model.CartItem
 
 class CartViewHolder(
