+++ app/src/main/java/woowacourse/shopping/feature/main/viewmodel/MainViewModelFactory.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.feature.main.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.data.InquiryHistoryRepository
+import woowacourse.shopping.data.inquiryhistory.InquiryHistoryRepository
 import woowacourse.shopping.data.cart.CartRepository
 import woowacourse.shopping.data.product.ProductRepository
 import java.lang.IllegalArgumentException
