+++ app/src/main/res/layout/activity_product_detail.xml
@@ -12,7 +12,7 @@
 
         <variable
             name="quantityControlListener"
-            type="woowacourse.shopping.feature.main.QuantityControlListener" />
+            type="woowacourse.shopping.common.QuantityControlListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
