+++ app/src/main/res/layout/button_quantity_control.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="quantityControlListener"
-            type="woowacourse.shopping.feature.main.QuantityControlListener" />
+            type="woowacourse.shopping.common.QuantityControlListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
