+++ app/src/main/res/layout/item_product.xml
@@ -18,7 +18,7 @@
 
         <variable
             name="quantityControlListener"
-            type="woowacourse.shopping.feature.main.QuantityControlListener" />
+            type="woowacourse.shopping.common.QuantityControlListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
